src/backend/requirements.txt
asgiref==3.8.1
black==25.1.0
Django==4.2.20
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==23.0.0
isort==6.0.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.8.0
pytz==2025.1
ruff==0.9.10
setuptools-76.0.0
sqlparse==0.5.3
